Function for verifying solution
Added function to verify a solution.
Users solution is stored as an Array of transition pairs(array). The required Passenger transitions defined in the model.ScenarioLevels.ScenarioSolutions are tested against the users solution, resulting in a PASS if all transitions for that passenger matches, or FAIL if one transition pair is not present in the users solution.